Build Qt 5.10 with openssl 1.1+

+++ libs/qt5/qtbase/qtbase.py
@@ -70,16 +70,16 @@ class subinfo(info.infoclass):
 
     def setDependencies(self):
         if CraftCore.settings.getboolean("Packager", "UseCache") and not CraftCore.settings.getboolean("QtSDK", "Enabled", False):
-            self.buildDependencies["dev-util/qtbinpatcher"] = "default"
-        self.runtimeDependencies["virtual/base"] = "default"
-        self.buildDependencies["dev-util/perl"] = "default"
-        self.buildDependencies["dev-util/winflexbison"] = "default"
+            self.buildDependencies["dev-util/qtbinpatcher"] = None
+        self.runtimeDependencies["virtual/base"] = None
+        self.buildDependencies["dev-util/perl"] = None
+        self.buildDependencies["dev-util/winflexbison"] = None
         if not self.options.buildStatic:
-            self.runtimeDependencies["win32libs/openssl"] = "default"
-            self.runtimeDependencies["win32libs/dbus"] = "default"
-            self.runtimeDependencies["binary/mysql"] = "default"
-            self.runtimeDependencies["win32libs/icu"] = "default"
-            self.runtimeDependencies["win32libs/zlib"] = "default"
+            self.runtimeDependencies["win32libs/openssl"] = None if CraftVersion(self.buildTarget) < CraftVersion("5.10") else "1.1"
+            self.runtimeDependencies["win32libs/dbus"] = None
+            self.runtimeDependencies["binary/mysql"] = None
+            self.runtimeDependencies["win32libs/icu"] = None
+            self.runtimeDependencies["win32libs/zlib"] = None
 
 
 class QtPackage(Qt5CorePackageBase):
@@ -98,12 +98,14 @@ class QtPackage(Qt5CorePackageBase):
             return False
         self.enterBuildDir()
         if OsUtils.isWin():
-            configure = os.path.join(self.sourceDir(), "configure.bat").replace("/", "\\")
-            if not os.path.exists(os.path.join(self.sourceDir(), ".gitignore")):  # force bootstrap of configure.exe
-                with open(os.path.join(self.sourceDir(), ".gitignore"), "wt+") as bootstrap:
-                    bootstrap.write("Force Bootstrap")
-                if os.path.exists(os.path.join(self.sourceDir(), "configure.exe")):
-                    os.remove(os.path.join(self.sourceDir(), "configure.exe"))
+            configure = OsUtils.toUnixPath(os.path.join(self.sourceDir(), "configure.bat"))
+            if self.qtVer < CraftVersion("5.10"):
+                # not needed anymore as we don't patch configure anymore
+                if not os.path.exists(os.path.join(self.sourceDir(), ".gitignore")):  # force bootstrap of configure.exe
+                    with open(os.path.join(self.sourceDir(), ".gitignore"), "wt+") as bootstrap:
+                        bootstrap.write("Force Bootstrap")
+                    if os.path.exists(os.path.join(self.sourceDir(), "configure.exe")):
+                        os.remove(os.path.join(self.sourceDir(), "configure.exe"))
         elif OsUtils.isUnix():
             configure = os.path.join(self.sourceDir(), "configure")
 
@@ -137,12 +139,15 @@ class QtPackage(Qt5CorePackageBase):
             os.path.join(CraftStandardDirs.craftRoot(), "include"), os.path.join(CraftStandardDirs.craftRoot(), "lib"))
             if self.subinfo.options.isActive("win32libs/openssl"):
                 command += " -openssl-linked "
+                if self.qtVer >= CraftVersion("5.10"):
+                    opensslIncDir = os.path.join(CraftCore.standardDirs.craftRoot(), "include", "openssl")
+                    command += f" OPENSSL_LIBS=\"-llibssl -llibcrypto\" OPENSSL_INCDIR=\"{opensslIncDir}\" "
             if self.subinfo.options.isActive("binary/mysql"):
                 command += " -plugin-sql-mysql "
             if self.subinfo.options.isActive("win32libs/dbus"):
                 command += " -qdbus -dbus-runtime -I \"%s\" -I \"%s\" " % (
-                os.path.join(CraftStandardDirs.craftRoot(), "include/dbus-1.0"),
-                os.path.join(CraftStandardDirs.craftRoot(), "lib/dbus-1.0/include"))
+                os.path.join(CraftStandardDirs.craftRoot(), "include", "dbus-1.0"),
+                os.path.join(CraftStandardDirs.craftRoot(), "lib", "dbus-1.0", "include"))
             if self.subinfo.options.isActive("win32libs/icu"):
                 command += " -icu "
             if self.subinfo.options.isActive("win32libs/zlib"):
@@ -158,7 +163,7 @@ class QtPackage(Qt5CorePackageBase):
         if (CraftCore.compiler.isMSVC() and CraftCore.compiler.isClang()) or OsUtils.isUnix() or self.supportsCCACHE:
             command += "-no-pch "
 
-        if CraftCore.compiler.isMinGW():
+        if CraftCore.compiler.isMinGW() and self.qtVer < CraftVersion("5.10"):
             command += """ "QMAKE_CXXFLAGS += -Wa,-mbig-obj" """
 
         return utils.system(command)
